refactor: 구조 수정
